requirements.txt
requests
python-slugify
cssselect
lxml